Moved Editor Factory Methods for creating Settings into the HierarchyDecorator.cs class to remove UnityEditor namespace dependencies
